requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
langchain-nvidia-ai-endpoints
python-dotenv
pydantic
duckduckgo-search